Move add_root_to_dict to phylo.utils
